Refactor: DB関連ユーティリティをutils/db.tsに移動し、パスエイリアスを使用するように修正
