+++ api/src/routes/employees/tests/delete_employee.test.ts
@@ -1,15 +1,15 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import deleteEmployeeApp from '../delete_employee';
-import { createTestEnv } from '../../../tests/helpers';
-import * as fetchModule from '../../../model/fetch';
-import * as updateModule from '../../../model/update';
+import { createTestEnv } from '@tests/helpers';
+import * as fetchModule from '@model/fetch';
+import * as updateModule from '@model/update';
 
 // getEmployeeById, deactivateEmployee, deleteEmployee をモック化
-vi.mock('../../../model/fetch', () => ({
+vi.mock('@model/fetch', () => ({
   getEmployeeById: vi.fn()
 }));
 
-vi.mock('../../../model/update', () => ({
+vi.mock('@model/update', () => ({
   deactivateEmployee: vi.fn(),
   deleteEmployee: vi.fn()
 }));
